[TASK] Rename service class
