+++ Classes/Service/FlashMessageGenerator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace WebVision\WvDeepltranslate\Service\Handler;
+namespace WebVision\WvDeepltranslate\Service;
 
 use Exception;
 use TYPO3\CMS\Core\Messaging\FlashMessage;
@@ -11,7 +11,7 @@ use TYPO3\CMS\Core\Messaging\FlashMessageService;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 
-class ExceptionHandlerService
+class FlashMessageGenerator
 {
     protected FlashMessageQueue $messageQueue;
